Refactors and FIXMEs
