+++ packages/online-editor/src/accounts/github/ConnectToGitHubSection.tsx
@@ -32,11 +32,11 @@ import { getGithubInstanceApiUrl } from "../../github/Hooks";
 import { useOnlineI18n } from "../../i18n";
 import { useCancelableEffect } from "@kie-tools-core/react-hooks/dist/useCancelableEffect";
 import { AccountsDispatchActionKind, AccountsSection, useAccounts, useAccountsDispatch } from "../AccountsContext";
-import { useAuthSessions, useAuthSessionsDispatch } from "../authSessions/AuthSessionsContext";
-import { AuthSessionDescriptionList } from "../authSessions/AuthSessionsList";
-import { GitAuthSession } from "../authSessions/AuthSessionApi";
+import { useAuthSessions, useAuthSessionsDispatch } from "../../authSessions/AuthSessionsContext";
+import { AuthSessionDescriptionList } from "../../authSessions/AuthSessionsList";
+import { GitAuthSession } from "../../authSessions/AuthSessionApi";
 import { PromiseStateStatus, usePromiseState } from "@kie-tools-core/react-hooks/dist/PromiseState";
-import { GitAuthProvider } from "../authProviders/AuthProvidersApi";
+import { GitAuthProvider } from "../../authProviders/AuthProvidersApi";
 
 export const GITHUB_OAUTH_TOKEN_SIZE = 40;
