+++ packages/online-editor/src/authSessions/AuthSessionApi.ts
