+++ packages/online-editor/src/authSessions/CompatibleAuthSessions.ts
