+++ packages/online-editor/src/devDeployments/DevDeploymentsConfirmDeployModal.tsx
@@ -23,7 +23,7 @@ import { useOnlineI18n } from "../i18n";
 import { WorkspaceFile } from "@kie-tools-core/workspaces-git-fs/dist/context/WorkspacesContext";
 import { useDevDeployments } from "./DevDeploymentsContext";
 import { useGlobalAlert } from "../alerts";
-import { useAuthSession } from "../accounts/authSessions/AuthSessionsContext";
+import { useAuthSession } from "../authSessions/AuthSessionsContext";
 
 interface Props {
   workspaceFile: WorkspaceFile;
