+++ packages/online-editor/src/home/UploadCard.tsx
@@ -30,7 +30,7 @@ import { UploadIcon } from "@patternfly/react-icons/dist/js/icons/upload-icon";
 import { Divider } from "@patternfly/react-core/dist/js/components/Divider";
 import { Split, SplitItem } from "@patternfly/react-core/dist/js/layouts/Split";
 import { LocalFile } from "@kie-tools-core/workspaces-git-fs/dist/worker/api/LocalFile";
-import { AUTH_SESSION_NONE } from "../accounts/authSessions/AuthSessionApi";
+import { AUTH_SESSION_NONE } from "../authSessions/AuthSessionApi";
 
 enum UploadType {
   NONE,
