+++ packages/online-editor/src/importFromUrl/ImportFromUrlHomePageCard.tsx
@@ -28,10 +28,10 @@ import * as React from "react";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { useHistory } from "react-router";
 import { AccountsDispatchActionKind, useAccountsDispatch } from "../accounts/AccountsContext";
-import { useAuthProviders } from "../accounts/authProviders/AuthProvidersContext";
-import { AUTH_SESSION_NONE } from "../accounts/authSessions/AuthSessionApi";
-import { useAuthSession, useAuthSessions } from "../accounts/authSessions/AuthSessionsContext";
-import { getCompatibleAuthSessionWithUrlDomain } from "../accounts/authSessions/CompatibleAuthSessions";
+import { useAuthProviders } from "../authProviders/AuthProvidersContext";
+import { AUTH_SESSION_NONE } from "../authSessions/AuthSessionApi";
+import { useAuthSession, useAuthSessions } from "../authSessions/AuthSessionsContext";
+import { getCompatibleAuthSessionWithUrlDomain } from "../authSessions/CompatibleAuthSessions";
 import { useRoutes } from "../navigation/Hooks";
 import { AdvancedImportModal, AdvancedImportModalRef } from "./AdvancedImportModalContent";
 import { isPotentiallyGit, useClonableUrl, useImportableUrl, useImportableUrlValidation } from "./ImportableUrlHooks";
