+++ packages/online-editor/src/importFromUrl/ImportSingleFileFromUrlForm.tsx
@@ -21,9 +21,9 @@ import { CheckCircleIcon } from "@patternfly/react-icons/dist/js/icons/check-cir
 import { ExclamationCircleIcon } from "@patternfly/react-icons/dist/js/icons/exclamation-circle-icon";
 import * as React from "react";
 import { FormEvent, useCallback, useMemo } from "react";
-import { AuthProviderGroup } from "../accounts/authProviders/AuthProvidersApi";
-import { AuthSessionSelect } from "../accounts/authSessions/AuthSessionSelect";
-import { gitAuthSessionSelectFilter } from "../accounts/authSessions/CompatibleAuthSessions";
+import { AuthProviderGroup } from "../authProviders/AuthProvidersApi";
+import { AuthSessionSelect } from "../authSessions/AuthSessionSelect";
+import { gitAuthSessionSelectFilter } from "../authSessions/CompatibleAuthSessions";
 import { ImportableUrl } from "./ImportableUrlHooks";
 
 export function ImportSingleFileFromUrlForm(props: {
