+++ packages/online-editor/src/importFromUrl/ImportableUrlHooks.tsx
@@ -25,10 +25,10 @@ import { extname } from "path";
 import * as React from "react";
 import { useCallback, useMemo } from "react";
 import { matchPath } from "react-router";
-import { AuthProviderIcon } from "../accounts/authProviders/AuthProviderIcon";
-import { useAuthProvider } from "../accounts/authProviders/AuthProvidersContext";
-import { AuthSession } from "../accounts/authSessions/AuthSessionApi";
-import { AuthInfo } from "../accounts/authSessions/AuthSessionsContext";
+import { AuthProviderIcon } from "../authProviders/AuthProviderIcon";
+import { useAuthProvider } from "../authProviders/AuthProvidersContext";
+import { AuthSession } from "../authSessions/AuthSessionApi";
+import { AuthInfo } from "../authSessions/AuthSessionsContext";
 import { useEditorEnvelopeLocator } from "../envelopeLocator/hooks/EditorEnvelopeLocatorContext";
 import { AdvancedImportModalRef } from "./AdvancedImportModalContent";
 import { getGitRefName, getGitRefType } from "../gitRefs/GitRefs";
